spaces

+++ contracts/mocks/Token.sol
@@ -6,9 +6,10 @@ import "@openzeppelin/contracts/access/Ownable.sol";
 
 contract Token is ERC20, Ownable {
   constructor(string memory symbol) ERC20(symbol, symbol) {
-    _mint(msg.sender, 10 ** 36);
+    _mint(msg.sender, 10**36);
   }
-  function mintTo(address account, uint amount) public onlyOwner {
+
+  function mintTo(address account, uint256 amount) public onlyOwner {
     _mint(account, amount);
   }
 }
